requirements.txt
python-dotenv==1.0.1
requests==2.32.3
aiogram==3.10.0
asyncpg==0.29.0
SQLAlchemy==2.0.32
alembic==1.13.2